refactor-home-page
type: refactor
- homepage moved to stateless component
 - moved state to app.js
